reports/requirements.txt
pandas>=2.2
openpyxl>=3.1
jinja2>=3.1
matplotlib>=3.8
python-dateutil>=2.9
gspread>=6.1
google-auth>=2.30
python-dotenv>=1.0
